PartyRecomender/requirements.txt
fastapi==0.68.0
tekore==3.7.1
uvicorn==0.15.0
pandas==1.3.2
flask==2.0.1
psycopg2-binary==2.9.1